fix cascading deletes to work with paranoid

+++ dashboard/app/models/lti_section.rb
@@ -22,9 +22,9 @@ class LtiSection < ApplicationRecord
   belongs_to :section
   validates :section_id, uniqueness: true
   validates :lms_section_id, presence: true
-  after_destroy :destroy_associated_section
+  before_destroy :soft_delete_associated_section
 
-  def destroy_associated_section
-    section.destroy
+  def soft_delete_associated_section
+    section.destroy unless section.deleted?
   end
 end
